Беру изменения

+++ English-Bot/MsgHandle.cs
@@ -164,14 +164,26 @@ namespace English_Bot
                     DayWords_Start(fromId);
                     return;
 
+                case "admin::getpicture":
+                    if (adminIDs.Contains(fromId))
+                    {
+                        foreach (var x in users[fromId].unLearnedWords)
+                            SendPicture(fromId, x);
+                    }
+                    else answer = ACCESS_IS_DENIED;
+                    return;
                 case "admin::getсommands":
                     if (adminIDs.Contains(fromId))
-                        answer = "getId, wantTest, getCommands, usersCount, forget_me";
+                        answer = "getId, wantTest, getCommands, usersCount, forget_me, getpicture";
                     else answer = ACCESS_IS_DENIED;
                     break;
                 case "admin::forget_me":
-                    SendMessage(fromId, "Я тебя забыл.");
-                    users.DeleteUser(fromId);
+                    if (adminIDs.Contains(fromId))
+                    {
+                        SendMessage(fromId, "Я тебя забыл.");
+                        users.DeleteUser(fromId);
+                    }
+                    else answer = ACCESS_IS_DENIED;
                     return;
                 case "admin::getid":
                     if (adminIDs.Contains(fromId))
@@ -329,8 +341,8 @@ namespace English_Bot
         {
             long id = (long)Idobj;
             var user = users[id];
-            SendMessage(id, "Выберите один из вариантов:",null, true);
-            SendMessage(id, "Лёгкий - выбор из предложенных вариантов ответов на тестировании.", null, true);
+            SendMessage(id, "Выберите один из вариантов:");
+            SendMessage(id, "Лёгкий - выбор из предложенных вариантов ответов на тестировании.");
             SendMessage(id, "Сложный - полностью самостоятельный перевод слов на тестировании.", null, true);
             var x = WaitWordFromUser_with_Comments(id, new string[] { "легкий", "сложный" }, 2,"Ладно, потом...");
             user.keyb = User.Main_Keyboard;
